+++ app/Console/Jobs/SendVerifyEmail.php
@@ -14,7 +14,7 @@ class SendVerifyEmail implements QueueInterface
     private $subject;
     private $max_attempts;
 
-    public function __construct(array $user, string $subject = "Verify your email", $max_attempts = 3)
+    public function __construct(array $user, string $subject = "Verify your email", $max_attempts = 300)
     {
         $this->user = $user;
         $this->subject = $subject;
@@ -28,13 +28,13 @@ class SendVerifyEmail implements QueueInterface
      */
     public function handle()
     {
-        if (env('APP_ENV') == 'local' || env('SEND_EMAIL_ON_LOCAL') != 'true') {
+        if (env('APP_ENV') == 'local' && env('SEND_EMAIL_ON_LOCAL') != 'true') {
             return $this->delete();
         }
         try {
-            logger(storage_path()."logs/email.log")->info("sending verification email to {$this->user['email']}");
+            logger(storage_path("logs/email.log"))->info("sending verification email to {$this->user['email']}");
 
-            if ($this->job['tries'] >= $this->max_attempts)
+            if ($this->job['tries'] > $this->max_attempts)
                 return $this->fail();
 
             if (!VerifyEmail::send($this->user['email'], $this->user['firstname'].' '.$this->user['lastname'], $this->subject, $this->user['email2fa_token']))
@@ -42,7 +42,7 @@ class SendVerifyEmail implements QueueInterface
 
             $this->delete();
         } catch (Exception $e) {
-            logger(storage_path()."logs/email.log")->error('Caught a ' . get_class($e) . ': ' . $e->getMessage(), $e->getTrace());
+            logger(storage_path("logs/email.log"))->error('Caught a ' . get_class($e) . ': ' . $e->getMessage(), $e->getTrace());
         }
     }
 }
